#10: Restructure * Removed unnecessary `usings` * Moved `*Ribbon.cs` files into own Ribbon folder
