+++ MarkdownOutlook/Ribbons/AppointmentRibbon.resx
